Create 7 October lin kSZ
The day when I tried to figure out why my lin kSZ limber plot was wrong..for the second time. I checked integrand (i.e. dD/dz) where D=l**2*Cl/2pi and checked that the shape was the same as in the modified gravity. Then I checked that the mu was not the problem by computing the integrand of the function in which I integrated over mu instead. I checked constants, units, order of magnitude.